# docs/reference/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "NURTW Membership and Vehicle Verification System",
    "version": "0.1.0",
    "description": "Internal administration API for the National Union of Road Transport Workers, Anambra State Council.\n\nA verification enquiry answered by this system confirms only that a matching NURTW record exists under the requested criteria. It is not evidence of ownership, roadworthiness, licensing, or insurance, and no response should be presented as such.\n\nThis document is generated from the running application: the route list is read from the dependency-injection container and request bodies from the same schemas the API validates against. It cannot describe a route that does not exist."
  },
  "servers": [
    {
      "url": "http://localhost:3001"
    }
  ],
  "components": {
    "securitySchemes": {
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "nurtw_session",
        "description": "Opaque session token issued by POST /auth/login and set as an httpOnly cookie. Revocation takes effect on the next request."
      }
    },
...
  },
  "security": [
    {
      "sessionCookie": []
    }
  ],
  "paths": {
    "/api/v1/auth/holders": {
      "get": {
        "summary": "List who currently holds a permission, and in what scope.",
        "description": "Answers across role bundles, per-user grants, and per-user revocations, with revocations applied. It exists because `vehicle.declare` is held so narrowly that it is worthless as a control if establishing who holds it requires reasoning through those three layers by hand.\n\n**Permission required:** `permission.read`.",
        "operationId": "AuthController_holders",
        "tags": [
          "Auth"
        ],
        "parameters": [
          {
            "name": "permission",
            "in": "query",
            "required": true,
            "description": "The permission code to enquire about, for example `vehicle.declare`.",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful response."
          },
          "400": {
            "description": "No permission code was supplied.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "401": {
            "description": "No valid session cookie was presented.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "The session does not hold permission.read in the relevant scope.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/auth/login": {
      "post": {
        "summary": "Open a session.",
        "description": "On success an opaque session token is set as an `httpOnly` cookie and is never returned in the body, so it cannot be recovered from a logged response or an XHR trace. An unknown account and an incorrect password produce byte-identical responses, and both take comparable time, so neither existence nor near-misses can be inferred.\n\n**Authentication:** none. This route is deliberately public.",
        "operationId": "AuthController_login",
        "tags": [
          "Auth"
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$schema": "https://json-schema.org/draft/2020-12/schema",
                "type": "object",
                "properties": {
                  "email": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 320
                  },
                  "password": {
                    "type": "string",
                    "minLength": 1,
                    "maxLength": 1024
                  }
                },
                "required": [
                  "email",
                  "password"
                ]
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": "Created."
          },
          "401": {
            "description": "The credentials were not accepted.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        },
        "security": []
      }
    },
...
    "/api/v1/auth/me": {
      "get": {
        "summary": "Describe the signed-in user and their effective permissions.",
        "description": "Returns each permission with the organisational scope in which it is held, so the dashboard can hide controls the user cannot use. Hiding a control is never the control itself — the guard remains the authority. Requires only `organisation.read`, which every internal role holds; this is a \"who am I\" endpoint, not a privileged one.\n\n**Permission required:** `organisation.read`.",
        "operationId": "AuthController_me",
        "tags": [
          "Auth"
        ],
        "responses": {
          "200": {
            "description": "Successful response."
          },
          "401": {
            "description": "No valid session cookie was presented.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          },
          "403": {
            "description": "The session does not hold organisation.read in the relevant scope.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": [
                    "error"
                  ],
                  "properties": {
                    "error": {
                      "type": "object",
                      "required": [
                        "code",
                        "message",
                        "requestId"
                      ],
                      "properties": {
                        "code": {
                          "type": "string"
                        },
                        "message": {
                          "type": "string",
                          "description": "Deliberately uninformative. Correlate the requestId with the server log for detail."
                        },
                        "requestId": {
                          "type": "string"
                        },
                        "details": {
                          "type": "array",
                          "description": "Field-level validation failures. Present only on a rejected request body, and describing only the request the caller sent.",
                          "items": {
                            "type": "object",
                            "required": [
                              "field",
                              "message"
                            ],
                            "properties": {
                              "field": {
                                "type": "string"
                              },
                              "message": {
                                "type": "string"
                              }
                            }
                          }
                        }
                      }
                    }
                  }
                }
              }
            }
          }
        }
      }
    },
...
  }
}